pyproject.toml
[tool.poetry]
name = "nedvizhka"
version = "0.1.0"
description = ""
authors = ["yahoryakubovich <officialegoryakubovich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
pillow = "^10.0.0"
django-allauth = "^0.54.0"
python-social-auth = {extras = ["yandex"], version = "^0.3.6"}
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
